+++ Vector3D.cpp
@@ -33,7 +33,7 @@ Vector3D Vector3D::Normal(const Vector3D & a)
 float Vector3D::Magnitude() { return sqrtf(x * x + y * y + z * z); }
 Vector3D Vector3D::Reflect(const Vector3D & v, const Vector3D & a) {
 	Vector3D n = Normal(a);
-	float co = -2 * (v.Dot(n) / (n.Magnitude() * n.Magnitude()));
+	float co = -2 * ((float)v.Dot(n) / (n.Magnitude() * n.Magnitude()));
 	Vector3D r = {};
 	r.x = v.x + co * n.x;
 	r.y = r.y + co * n.y;
